fix(packages/excalidraw): consolidate bounds checks (#11275)
